ai_filter/requirements.txt
ollama
pyyaml
requests
beautifulsoup4
PyYAML
